% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDotPlot.R
\name{EnrichDotPlot}
\alias{EnrichDotPlot}
\title{Enrichment Dot Plot}
\usage{
EnrichDotPlot(
  enrich.list,
  top = 3,
  only.top = TRUE,
  color = paletteer_dynamic("cartography::red.pal", 20)[10:20]
)
}
\arguments{
\item{enrich.list}{A list of enrichment results generated by \code{\link{PGEnricher}}}

\item{top}{The number of top enriched terms to be shown for each program.}

\item{only.top}{A logical value indicating whether to plot only the top enriched terms (TRUE) or all enriched terms that are part of the top enriched pathways (FALSE).}

\item{color}{A vector of colors to be used in the dot plot. The length of the vector should be at least the number of terms being plotted. The default is \code{paletteer::paletteer_dynamic("cartography::red.pal", 20)[10:20]}.}
}
\value{
A ggplot object representing the enrichment dot plot.
}
\description{
Visualize the enrichment results of transciptome programs using a dot plot. The function takes a list of enrichment results generated by \code{\link{PGEnricher}} as input and plots the top enriched terms for each program by \code{\link{ggplot2}}.
}
\seealso{
\code{\link[ggplot2]{ggplot}} for customizing the plot appearance.
}
