% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRPlot.R
\name{IQRPlot}
\alias{IQRPlot}
\title{IQRPlot: Generate Boxplots with IQR and Median Filter}
\usage{
IQRPlot(WH.list, IQR.cut = 0.1, median.cut = 0, grid = TRUE, ncol = NULL)
}
\arguments{
\item{WH.list}{A nested list containing NMF results for each individual, generated from the \code{\link{RunNMF}} function. The structure of WH.list should be:

\itemize{
 \item\code{sample_1}
 \itemize{
 \item\code{W} Normalized W matrix from NMF
 \item\code{H} Normalized H matrix from NMF
}
\item\code{sample_2}
 \itemize{
 \item\code{W} Normalized W matrix from NMF
 \item\code{H} Normalized H matrix from NMF
}
...
\item\code{sample_n}
}}

\item{IQR.cut}{The cutoff value for IQR. Programs with IQR above this value will be kept in the plot.}

\item{median.cut}{The cutoff value for median. Programs with median above this value will be kept in the plot.}

\item{grid}{A logical value indicating whether to arrange the plots of samples in a grid (TRUE) or return a list of ggplot objects (FALSE).}

\item{ncol}{Determines the number of columns when arranging the plots in a grid. If NULL, it will be automatically determined based on the number of samples.}
}
\value{
A girded ggplot object or a list of ggplot objects containing boxplot for each program with IQR and median filtering.
}
\description{
The IQRPlot function is used to generate boxplots for each program based on the IQR (Interquartile Range) and median filter. It calculates the IQR and median usage of each program from the NMF (Non-Negative Matrix Factorization) results and filters out programs with low IQR and median values. The higher IQR indicates a larger range of usage, which is useful for quality control.
}
\examples{
\dontrun{
    result <- IQRPlot(WH.list, IQR.cut = 0.1, median.cut = 0, grid = TRUE)
}

}
\keyword{NMF}
\keyword{boxplot}
\keyword{result}
\keyword{visualization}
