% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ls_gs_sciNMF}
\alias{ls_gs_sciNMF}
\title{List of sciNMF built-in gene sets data.frame}
\format{
A list with 5 TERM TO GENE mapping data.frames and their sources:
\itemize{
  \item\code{CellMarker}: CellMarker_Augmented_2021.txt
  \item\code{GO_BP}: GO_Biological_Process_2023.txt
  \item\code{GO_CC}: GO_Cellular_Component_2023.txt
  \item\code{GO_MF}: GO_Molecular_Function_2023.txt
  \item\code{Hallmark}: MSigDB_Hallmark_2020.txt
}
}
\source{
\url{https://maayanlab.cloud/Enrichr/#libraries}
}
\usage{
ls_gs_sciNMF[["GO_BP"]]
}
\description{
A list of annotation of TERM TO GENE mapping data.frames, which have 2 columns with term and gene. The data.frames are generated by \code{\link[gson]{read.gmt}}
}
\examples{
data("ls_gs_sciNMF")
}
\seealso{
\code{\link[gson]{read.gmt}} for gene.set data.frame generation.
}
\keyword{datasets}
