% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaProgram.R
\name{MetaProgram}
\alias{MetaProgram}
\title{Construct Meta Programs from NMF and Clustering Results}
\usage{
MetaProgram(WH.list, cluster.result, top = 50, key = "MP_", only.gene = TRUE)
}
\arguments{
\item{WH.list}{A nested list containing NMF results for each individual, generated from the \code{\link{RunNMF}} function. The structure of WH.list should be:

\itemize{
 \item\code{sample_1}
 \itemize{
 \item\code{W} Normalized W matrix from NMF
 \item\code{H} Normalized H matrix from NMF
}
\item\code{sample_2}
 \itemize{
 \item\code{W} Normalized W matrix from NMF
 \item\code{H} Normalized H matrix from NMF
}
...
\item\code{sample_n}
}}

\item{cluster.result}{The result of \code{\link{OvlpHeatmap}} function contains cluster reults for each program. Or a named vector of cluster result.}

\item{top}{The number of top genes to extract for each meta-program. Default is 50.}

\item{key}{A prefix string for the names of extracted meta-programs. Default is 'MP_'.}

\item{only.gene}{Logical, if TRUE, only gene names are returned; if FALSE, named gene scores are returned. Default is TRUE.}
}
\value{
A list of extracted meta-programs.
Each element of the list represents a meta-program and contains either gene names or gene scores (if only.gene = FALSE).
}
\description{
This function extracts meta-programs from the results of Non-negative Matrix Factorization (NMF) and cluster relationship.
}
\details{
The function extracts meta-programs from the NMF results by aggregating the results from multiple samples.
It first aligns the genes across all samples and then computes the average expression profile for each meta-program.
The top genes with the highest average expression scores are selected and returned.
}
\seealso{
\code{\link{RunNMF}} for the function to perform NMF analysis.
}
