% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPCoxph.R
\name{MPCoxph}
\alias{MPCoxph}
\title{Cox Proportional Hazards Regression for MetaPrograms}
\usage{
MPCoxph(
  mat.exp,
  cli,
  gene.list,
  time = "OS.time",
  event = "OS.status",
  group.by = NULL,
  min.sample = 15,
  return.df = FALSE,
  xlab = "MetaProgram",
  ylab = "Group",
  titile = "Univariate Cox Regression",
  color.asterisks = "white",
  show.ns = FALSE,
  score.method = "ssgsea",
  kcdf = "Gaussian"
)
}
\arguments{
\item{mat.exp}{The gene expression matrix with genes as rows and samples as columns.}

\item{cli}{A data.frame contains clinical data with survival information, including time-to-event (time) and event status (event).}

\item{gene.list}{A list of gene sets (meta-programs) with each element containing the genes in a meta-program.}

\item{time}{The column name in the \code{cli} data.frame that represents the time-to-event (survival time). Default is "OS.time"}

\item{event}{The column name in the \code{cli} data.frame that represents the event status (event occurrence).
This column should contain binary values (0 or 1), where 0 indicates that the event has not occurred (alive) and 1 indicates that the event has occurred (dead).}

\item{group.by}{A character vector specifying the column names in the \code{cli} data.frame by which the samples will be split for subgroup analysis. If NULL, all samples will be treated as one group.}

\item{min.sample}{The minimum number of samples required in each subgroup for Cox regression analysis. Default is 15.}

\item{return.df}{A logical value indicating whether to return the result as a data frame (TRUE) or as a ggplot object (FALSE).}

\item{xlab}{The label for the x-axis of the plot. Default is "Program".}

\item{ylab}{The label for the y-axis of the plot. Default is "Group".}

\item{titile}{The title of the plot. Default is "Univariate Cox Regression".}

\item{color.asterisks}{The color for indicating the significance level of each meta-program.}

\item{show.ns}{A logical value indicating whether to show ns (no significant).}

\item{score.method}{The method used to calculate the meta-program score, which can be one of "ssgsea", "gsva", "plage", or "average".}

\item{kcdf}{Character string denoting the kernel to use during the non-parametric estimation of the cumulative distribution function of expression levels across samples when method="gsva". By default, kcdf="Gaussian". More details are in \code{GSVA::\link[GSVA]{gsva}}.}
}
\value{
A ggplot object representing the Cox proportional hazards regression results for meta-programs.
The significance level of each meta-program is indicated by the number of asterisks (*) as follows:
\itemize{ \item\code{ns}: p > 0.05 \item\code{*}: p <= 0.05 \item \code{**}: p <= 0.01 \item \code{***}: p <= 0.001 \item \code{****}:  p <= 0.0001 }
}
\description{
Perform Cox proportional hazards regression for a set of meta-programs (gene sets) on survival data.
}
\seealso{
\code{\link[ggplot2]{ggplot}} for customizing the plot appearance.

\code{\link[GSVA]{gsva}} for scoring the samples by the programs.

\code{\link[survival]{coxph}} for univariate cox regression
}
