% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OvlpHeatmap.R
\name{OvlpHeatmap}
\alias{OvlpHeatmap}
\title{Visualize the overlap result by heatmap.}
\usage{
OvlpHeatmap(
  ovlp,
  cut.num = NA,
  border = "grey",
  method.clustering = "ward.D2",
  distance.clustering = "correlation",
  annotation = NA,
  color.annotation = NA,
  cluster.name = "MetaProgram",
  cluster.key = "MP_",
  name = "Share Genes",
  show.colnames = FALSE,
  breaks = c(0:25),
  color = c("white", rev(as.character(paletteer::paletteer_c("viridis::magma",
    32))[8:32])),
  ...
)
}
\arguments{
\item{ovlp}{A square matrix generated by \code{\link{OverlapMat}}, representing the overlap scores between gene sets.}

\item{cut.num}{Number of clusters to cut the heatmap, if provided, the heatmap will be clustered based on the given number of clusters and
return will be a list contained a Heatmap and cluster result.}

\item{border}{Color of the heatmap borders.}

\item{method.clustering}{The clustering method. Possible values
are all the methods supported by \code{\link[stats]{hclust}} such as \code{"ward.D2"}, \code{"single"}, \code{"average"}, etc. Default is \code{"ward.D2"}.}

\item{distance.clustering}{The distance measure used for clustering. Possible values are
\code{"correlation"} for Pearson correlation, \code{"Jaccard"} for Jaccard similarity and all the distances supported by \code{\link[stats]{dist}}, such as \code{"euclidean"}, etc.
Default is \code{"correlation"}. If this parameter is set as \code{"Jaccard"}, the \code{Jaccard} parameter in \code{\link{OverlapMat}} should be set as \code{TRUE}.}

\item{annotation}{A data.frame containing additional annotations to be shown as row annotations on the heatmap, default is NA.}

\item{color.annotation}{A list of colors to be used for the annotations, default is NA. The names of this list should match with the colnames of annotation data.frame}

\item{cluster.name}{The name of the cluster column in the annotation data.frame, default is 'MetaProgram'.}

\item{cluster.key}{The prefix for the cluster names, default is 'MP_'.}

\item{name}{The name of the heatmap legend, default is 'Share Genes'.}

\item{show.colnames}{A logical value indicates whether column names are be shown.}

\item{breaks}{A numeric vector indicates the breakpoints for the color scale of the heatmap.}

\item{color}{The colors to be used for the heatmap.}

\item{...}{Additional arguments to be passed to \code{ComplexHeatmap::\link[ComplexHeatmap]{pheatmap}}.}
}
\value{
When the \code{cut.num} is provided, a list contains 2 element: Heatmap and Cluster will be returned, and this list can be as the input of \code{\link{MetaProgram}}. Otherwise, only a Heatmap will be returned.
\describe{
  \item{Heatmap}{A Heatmap object generated by the \code{ComplexHeatmap::\link[ComplexHeatmap]{pheatmap}} function.}
  \item{Cluster}{If \code{cut.num} is provided, a named numeric vector containing the cluster assignments for each row in the heatmap.}
}
}
\description{
This function generates a heatmap to visualize the overlap result between different gene sets.
}
\examples{
# Sample transcriptome programs
set.seed(123)
ls_pg <- lapply(1:50, function(i){
  paste0('g',sample(1:100,20))
})

names(ls_pg) <- paste0('PG_',1:50)

# Generate the ovelap matrix
ovlp1 <- OverlapMat(ls_pg)
# Generate cluster plot
hm1 <- OvlpHeatmap(ovlp1, cut.num = 3)
print(hm1)

# Use the Jaccard similarity to cluster
ovlp2 <- OverlapMat(ls_pg, Jaccard = TRUE)
hm2 <- OvlpHeatmap(ovlp2, cut.num = 3, breaks = c(0,1))
print(hm2)

}
\seealso{
\code{ComplexHeatmap::\link[ComplexHeatmap]{pheatmap}} to customize the heatmap appearance.

\code{stats::\link[stats]{hclust}} for hierarchical clustering based on the specified number of clusters.
}
