% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustProgram.R
\name{RobustProgram}
\alias{RobustProgram}
\title{Identify Robust Programs using modified Nature 2023 Method}
\usage{
RobustProgram(
  WH.list,
  top = 50,
  IQR.cut = 0.1,
  median.cut = 0,
  intra.min = 35,
  intra.rep = 1,
  inter.filter = TRUE,
  inter.min = 10,
  inter.rep = 1,
  intra.max = 10
)
}
\arguments{
\item{WH.list}{A nested list containing NMF results for each individual, generated from the \code{\link{RunNMF}} function. The structure of WH.list should be:

\itemize{
 \item\code{sample_1}
 \itemize{
 \item\code{W} W matrix from NMF
 \item\code{H} H matrix from NMF
}
\item\code{sample_2}
 \itemize{
 \item\code{W} W matrix from NMF
 \item\code{H} H matrix from NMF
}
...
\item\code{sample_n}
}}

\item{top}{The number of top genes to extract for each robust program. Default is 50.}

\item{IQR.cut}{The threshold for the interquartile range (IQR) of gene usage to filter programs during QC. Default is 0.1.}

\item{median.cut}{The threshold for the median gene usage to filter programs during QC. Default is 0.02.}

\item{intra.min}{The threshold to identify similar programs within an individual
Number of overlap genes greater than or equal to \code{intra.min} are considered similar within the same individual
Default is 35}

\item{intra.rep}{The minimum number of similar programs. If a program has less than \code{intra.rep} similar programs in an individual,
it will be removed. Default is 1}

\item{inter.filter}{Logical value indicates whether filter the programs based on inter-tumor similarities. Default is TRUE}

\item{inter.min}{The threshold to identify robust programs across individuals
Number of overlap genes greater than or equal to \code{inter.min} are considered similar across individuals
Default is 10}

\item{inter.rep}{The minimum number of similar programs across other individuals to be considered as a robust program after intra-tumor filtering.
Default is 1}

\item{intra.max}{The threshold to identify redundant programs within an individual
Number of overlap genes greater than or equal to \code{intra.max} are considered redundant within the same individual
Default is 10}
}
\value{
A list containing robust gene programs for each individual
Each element of the list represents a program and contains \code{top} genes with gene names.
}
\description{
This function identifies robust gene programs using a modified method proposed in the Nature 2023 paper.
The method aims to remove redundant programs that are highly similar within the same patient.
}
\details{
The function performs Quality Control (QC) on the programs based on the Interquartile Range (IQR) and median usage (the intra-rank-normalized H matrix).
Programs with low IQR and median usage are filtered out.

Following QC, it removes programs that have fewer than \code{intra.rep} similar programs (number of overlapping genes between two programs >= \code{intra.min}) within the same individual (for selecting robust programs).

Subsequently, it filters out programs which have fewer than \code{inter.rep} similar programs (number of overlapping genes between two programs >= \code{inter.min}) across individuals.

Finally, it further filters out redundant programs, selecting only programs that have an intersection smaller than \code{intra.max} with previously selected programs within an individual.

The remaining programs are considered robust and are returned as the final result.
}
\references{
Gavish, Avishai et al. “Hallmarks of transcriptional intratumour heterogeneity across a thousand tumours.” Nature vol. 618,7965 (2023): 598-606. \url{https://doi.org/10.1038/s41586-023-06130-4}
}
\seealso{
\code{\link{RunNMF}} for the function to perform NMF analysis.
}
