% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SrtObj}
\alias{SrtObj}
\title{A seurat object with 400 cells and 13,642 features}
\format{
\describe{
  \item{Seurat object}{
    13642 features across 400 samples within 1 assay
    Active assay: RNA (13642 features, 0 variable features)
  }
}
}
\usage{
SrtObj
}
\description{
This object is used to test the package installation. It contains 400 leukemic cells from 4 AML patients in our study.
}
\examples{
data("SrtObj")
}
\keyword{datasets}
