% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterMetricsPlot.R
\name{ClusterMetricsPlot}
\alias{ClusterMetricsPlot}
\title{Cluster Metrics Plot}
\usage{
ClusterMetricsPlot(
  mat.ovlp,
  num.clusters = 2:20,
  ncol = 3,
  distance.clustering = "Intersection",
  max.intersect = 50,
  method.clustering = "ward.D2"
)
}
\arguments{
\item{mat.ovlp}{A symmetric numeric matrix generated by \code{\link{OverlapMat}}.}

\item{num.clusters}{The number range of clusters used for clustering analysis. The function will
calculate evaluation metrics for each specified number of clusters. Default is 2:20.}

\item{ncol}{The number of columns for arranging subplots in the plot area. Default is 3.}

\item{distance.clustering}{The distance measure used for clustering. Possible values are \code{"Intersection"} for intersection between two programs as similarity,
\code{"Jaccard"} for Jaccard similarity, \code{"correlation"} for Pearson correlation and all the distances supported by \code{\link[stats]{dist}}, such as \code{"euclidean"}, etc.
Default is \code{"Intersection"}. If this parameter is set as \code{"Jaccard"}, the \code{Jaccard} parameter in \code{\link{OverlapMat}} should be set as \code{TRUE}.}

\item{max.intersect}{When the \code{"distance.clustering"} is set to \code{"Intersection"}, the distance between two programs will be calculated as the difference of \code{max.intersect} and \code{intersection}. Default is 50.}

\item{method.clustering}{The clustering method. Possible values are all the methods supported by \code{\link[stats]{hclust}},
such as \code{"ward.D2"}, \code{"single"}, \code{"average"}, etc. Default is \code{"ward.D2"}.}
}
\value{
Returns a plot displaying cluster metrics.
 The plot includes lines and scatter points representing clustering evaluation metrics for different numbers of clusters.
}
\description{
This function is used to plot the clustering results of provided different numbers of clusters,
It calculates and visualizes key clustering evaluation metrics to aid in the assessment of clustering performance.
}
\examples{

# Sample transcriptional programs
set.seed(123)
ls_pg <- lapply(1:50, function(i) {
  paste0("g", sample(1:1000, 50))
})

# Generate the ovelap matrix
ovlp <- OverlapMat(ls_pg)
# Generate cluster plot
cluster_plot <- ClusterMetricsPlot(ovlp, num.clusters = 2:10, ncol = 2)

# Display the cluster plot
print(cluster_plot)

}
\seealso{
\code{\link{OverlapMat}}, \code{\link[stats]{hclust}}, \code{\link[stats]{dist}},  \code{\link[fpc]{cqcluster.stats}}
}
