% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDotPlot.R
\name{EnrichDotPlot}
\alias{EnrichDotPlot}
\title{Enrichment Dot Plot}
\usage{
EnrichDotPlot(
  enrich.list,
  top = 3,
  only.top = TRUE,
  title = "Enriched GeneSets",
  color = paletteer_dynamic("cartography::red.pal", 20)[10:20]
)
}
\arguments{
\item{enrich.list}{A list of enrichment results generated by \code{\link{PGEnricher}}}

\item{top}{The number of top enriched terms to be shown for each program.}

\item{only.top}{A logical value indicating whether to show only the top enriched terms (TRUE) for each program or show all enriched terms that belong to the union of top enriched pathways (FALSE).}

\item{title}{The title of the plot, default is 'Enriched GeneSets'}

\item{color}{A vector of colors to be used in the dot plot.}
}
\value{
A ggplot object representing the enrichment results.
}
\description{
Visualize the enrichment results of transcriptional programs using a dot plot. The function takes a list of enrichment results generated by \code{\link{PGEnricher}} as input and plots the top enriched terms for each program by \code{\link{ggplot2}}.
}
\seealso{
\code{\link[ggplot2]{ggplot}} for customizing the plot appearance.
}
