% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRPlot.R
\name{IQRPlot}
\alias{IQRPlot}
\title{IQRPlot: Boxplots with IQR and Median Filter}
\usage{
IQRPlot(
  WH.list,
  IQR.cut = 0.1,
  median.cut = 0.02,
  grid = TRUE,
  ncol = NULL,
  align = c("hv", "h", "v", "none")
)
}
\arguments{
\item{WH.list}{A nested list containing NMF results for each individual, generated by the \code{\link{RunNMF}} function.
 The structure of WH.list should be:
\itemize{
 \item\code{sample_1}
 \itemize{
 \item\code{W} W matrix from NMF
 \item\code{H} H matrix from NMF
}
\item\code{sample_2}
 \itemize{
 \item\code{W} W matrix from NMF
 \item\code{H} H matrix from NMF
}
...
\item\code{sample_n}
}}

\item{IQR.cut}{The threshold for IQR. Programs with IQR above this value will be kept in the plot, default is 0.1.}

\item{median.cut}{The threshold for median. Programs with median above this value will be kept in the plot, default is 0.02.}

\item{grid}{A logical value indicating whether to arrange the plots of samples in a grid (TRUE) or return a list of ggplot objects (FALSE).}

\item{ncol}{Determines the number of columns when arranging the plots in a grid. If NULL, it will be automatically determined based on the number of samples.}

\item{align}{(Optional) Specifies whether graphs in the grid should be horizontally ("h") or vertically ("v") aligned. Options are "none" (default), "hv" (align in both directions), "h", and "v".}
}
\value{
A girded ggplot object or a list of ggplot objects containing boxplots for each program with IQR and median filtering.
}
\description{
The IQRPlot function is used to generate boxplots for each program based on the Interquartile Range (IQR) and median filter. 
The IQR and median usage of each program is calculated from normalized H matrices, part of the the Non-Negative Matrix Factorization (NMF) results.
}
\details{
First, all the program usages in the same NMF rank for each cell will be normalized to 1. Then the IQR and median usage of each program will be calculated for quality control.
The programs with IQR and median usage below the cutoff will be labeled as 'Remove'
}
\examples{
\dontrun{
    result <- IQRPlot(WH.list, IQR.cut = 0.1, median.cut = 0.02, grid = TRUE)
}

}
\keyword{NMF}
\keyword{boxplot}
\keyword{result}
\keyword{visualization}
