% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaProgram.R
\name{MetaProgram}
\alias{MetaProgram}
\title{MetaProgram Generation}
\usage{
MetaProgram(
  WH.list,
  cluster.result,
  top = 50,
  min.size.MP = 30,
  key = "MP_",
  only.gene = TRUE,
  keep.rep.gene = TRUE,
  heatmap = TRUE,
  show.gap = TRUE,
  border = "grey",
  annotation = NULL,
  color.annotation = NULL,
  cluster.name = "MetaProgram",
  name = "Share Genes",
  show.colnames = FALSE,
  show.rownames = TRUE,
  show.anno = TRUE,
  breaks = 0:25,
  color = c("white", rev(as.character(paletteer::paletteer_c("viridis::magma",
    32))[8:32])),
  ...
)
}
\arguments{
\item{WH.list}{The NMF result list generated by \code{\link{RunNMF}}}

\item{cluster.result}{A vector generated by \code{\link{ClusterPG}} and it indicates the cluster assignment for each program.}

\item{top}{Number of top genes to be considered.}

\item{min.size.MP}{Minimum number of genes contained in a MetaProgram (MP), the MP without enough unique genes will be removed.}

\item{key}{A prefix to be added to the names of generated MPs.}

\item{only.gene}{Logical, indicating whether to include only genes or both genes and their coefficients.}

\item{keep.rep.gene}{Logical, indicating whether to keep redundant genes.}

\item{heatmap}{Logical, indicating whether to generate a heatmap.}

\item{show.gap}{Logical, indicating whether to show gaps in the heatmap.}

\item{border}{Color for the border in the heatmap.}

\item{annotation}{Additional annotation to be added to the heatmap. It should be a data.frame, each row represents a robust program contained in cluster.result,
each column represents an annotation item.}

\item{color.annotation}{A list object contain the annotation color, the name of each element should match the colnames of annotion data.frame 
and each element is a named color vector.}

\item{cluster.name}{Name for the clustering annotation (for heatmap metaprogram annotation name, defalut is MetaProgram).}

\item{name}{Name for the heatmap.}

\item{show.colnames}{Logical, indicating whether to show column names in the heatmap.}

\item{show.rownames}{Logical, indicating whether to show row names in the heatmap.}

\item{show.anno}{Logical, indicating whether to show annotation in the heatmap.}

\item{breaks}{Breaks for the heatmap color scale.}

\item{color}{Color palette for the heatmap.}

\item{...}{Additional arguments to be passed to the \code{ComplexHeatmap::\link[ComplexHeatmap]{pheatmap}} function.}
}
\value{
A list containing MetaProgram information and, if specified, a heatmap.
}
\description{
This function generates MetaPrograms and the heatmap of cluster result (optional).
}
\details{
The function including 3 steps:  

Step1: Generate the MetaProgram from cluster result. For each MP, the coefficient of genes are calculated 
from average weight (W matrix of NMF result) of all the programs assign in this MP, and the top genes will 
be selected as the MP signature. 
 
Step2: Remove redundant genes across all MPs, assign the overlap genes by the min rank then max coef across MPs. 
Then MPs contain no enough genes (\code{min.size.MP}) are removed. 
 
Step3: Draw the cluster heatmap of remaining MPs.
}
\seealso{
\code{ComplexHeatmap::\link[ComplexHeatmap]{pheatmap}}, \code{\link{RunNMF}}, \code{\link{ClusterPG}}
}
